Disambiguate to_r and to_cents methods in Tonal::Step
The Tonal::Step#to_r and Tonal::Step#to_cents are ambiguous. Are they returning
the values for the step or the values for the ratio from which the step is
derived?

To avoid the ambiguity, this commit adds the Tonal::Step instance method pairs:
step_to_r/ratio_to_r and step_to_cents/ratio_to_cents.

Additionally, this commit:

* Changes Tonal::Step#to_r and Tonal::Step#to_cents to be aliases of step_to_r
  and step_to_cents respectively. This change breaks backward compatibility.

* Adds Tonal::Step#efficiency and changes Tonal::Ratio#efficiency to use it.
  Note, the Tonal::Ratio value is the negative of the Tonal::Step value.

* Fixes a bug in Tonal::ReducedRatio#invert! where the returning instance wasn't
  being reduced after the antecedent and consequent were swapped.