Fix Tonal::Interval API
The upper/lower arguments to Tonal::Interval were not being applied consistently
everywhere. The upper ratio used in the computation should be the first argument
to the method, the lower ratio is the second argument.

This change makes the upper/lower arguments passed to Tonal::Interval consistent
in tests and in the call made from Tonal::ReducedRatio.