Add Numeric#decimal_power
Finding the decmial power of a number is one way of estimating its complexity
and also handy for estimating how to format it in presentations.

This commit adds the #decimal_power to the Numeric class.

It also adds the #to_a method to the Approximation::Set class, which is
delegated to the Set instance contained in the class.